Move celery & update periodic task signature
* Moving the celery configuration to the project layer
* Updating the periodic task signature...this needed a distinct task signature rather than calling the class directly
